+++ app/login/page.tsx
@@ -1,5 +1,10 @@
 import AuthForm from "@/components/auth-form";
+import { Suspense } from "react";
 
 export default function LoginPage() {
-  return <AuthForm mode="login" />;
+  return (
+    <Suspense fallback={<div></div>}>
+      <AuthForm mode="login" />
+    </Suspense>
+  );
 }
