+++ app/register/page.tsx
@@ -1,5 +1,10 @@
 import AuthForm from "@/components/auth-form";
+import { Suspense } from "react";
 
 export default function RegisterPage() {
-  return <AuthForm mode="register" />;
+  return (
+    <Suspense fallback={<div></div>}>
+      <AuthForm mode="register" />
+    </Suspense>
+  );
 }
