--- supabase/.gitignore
@@ -1,3 +1,0 @@
-# Supabase
-.branches
-.temp
