--- supabase/functions/.vscode/settings.json
@@ -1,5 +1,0 @@
-{
-    "deno.enable": true,
-    "deno.unstable": true,
-    "deno.importMap": "./import_map.json"
-}
