--- supabase/functions/_utils/stripe.ts
@@ -1,6 +1,0 @@
-import Stripe from "stripe";
-
-export const stripe = new Stripe(Deno.env.get("STRIPE_SECRET_KEY") ?? "", {
-  httpClient: Stripe.createFetchHttpClient(),
-  apiVersion: "2022-11-15",
-});
